+++ src/pages/SecurePage/index.js
@@ -2,11 +2,11 @@ import React, { Fragment } from 'react'
 import { Observer } from 'mobx-react-lite'
 import { Switch } from 'antd-mobile'
 import globalStore from 'global-state'
-import { useProvider } from 'contexts/router'
+import { useRouterContext } from 'contexts/router'
 import { useNaviContext } from 'contexts/navi'
 
 export default function SecurePage() {
-  const router = useProvider()
+  const router = useRouterContext()
   const Navi = useNaviContext()
   return <Observer>{() => (
     <Fragment>
@@ -15,6 +15,10 @@ export default function SecurePage() {
         <div className="dd-common-alignside" style={{ padding: '15px 20px', backgroundColor: 'white' }}>
           密码锁
         <Switch checked={globalStore.app.config.isLockerOpen} onChange={checked => globalStore.app.setLocker(checked)} />
+        </div>
+        <div className="dd-common-alignside" style={{ padding: '15px 20px', backgroundColor: 'white' }}>
+          调试
+        <Switch checked={globalStore.app.showDebug} onChange={checked => globalStore.app.toggleDebug()} />
         </div>
       </div>
     </Fragment>