+++ src/utils/utils.js
@@ -1,13 +1,13 @@
-import * as bowser from 'bowser'
+// import * as bowser from 'bowser'
 
 export function isDeskTop(dev) {
     return true
-    const browser = bowser.getParser(window.navigator.userAgent)
-    const isChromeApp = (window.matchMedia('(display-mode: standalone)').matches)
-    const isIosApp = window.navigator.standalone === true
-    const isDev = dev === undefined ? process.env.NODE_ENV === 'development' : dev
-    const platformType = browser.getPlatformType()
-    return isDev || (platformType === 'mobile' && (isChromeApp || isIosApp)) ? true : false
+    // const browser = bowser.getParser(window.navigator.userAgent)
+    // const isChromeApp = (window.matchMedia('(display-mode: standalone)').matches)
+    // const isIosApp = window.navigator.standalone === true
+    // const isDev = dev === undefined ? process.env.NODE_ENV === 'development' : dev
+    // const platformType = browser.getPlatformType()
+    // return isDev || (platformType === 'mobile' && (isChromeApp || isIosApp)) ? true : false
 }
 
 export function stringfyQuery(query = {}) {